requirements-ms-swift.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.7.0
bitsandbytes>=0.41.0
deepspeed>=0.12.0
fire>=0.5.0
ms-swift @ git+https://github.com/modelscope/ms-swift.git@main
